Add API for setting the amperage limit
This adds an api to the Client to allow for setting the new amperage
limit. It also exposes the available values for the limit.

The goal is to use this information to add a new Number entity to the HA
chargepoint integration (see:
https://github.com/mbillow/ha-chargepoint/issues/35)

API was discovered by sniffing packets from the ChargePoint app. Besides
the included unit tests this was also tested with my local device and
appears to work.